- extract ext lib into its components - add setting: auto enable gps on startup
