+++ os/config.lua
@@ -13,13 +13,14 @@ return {
   },
   settings = {
     refresh_delay = 0.5,
-    peripheral_module_ids = {},
-    display_name = "Unknown",
-    night_hour = 19,
     day_hour = 5,
-    max_msg_archive_file_size = 5120,
+    night_hour = 19,
     hours_off_set_from_utc = 0,
-    max_msg_archive_files = 10,
+    display_name = "Unknown",
     gps_enabled = true,
+    max_msg_archive_file_size = 5120,
+    enable_gps_onstartup = true,
+    peripheral_module_ids = {},
+    max_msg_archive_files = 10,
   },
 }
